requirements.txt
asgiref==3.7.2
Django==4.1.13
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.2.4
typing_extensions==4.8.0
uritemplate==4.1.1
